requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-appconf==1.0.5
django-imagekit==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
httpie==2.6.0
idna==3.3
mysqlclient==2.0.3
oauthlib==3.1.1
pilkit==2.0
Pillow==8.4.0
pycparser==2.20
Pygments==2.10.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
